requirements.txt
# Discord Bot
discord.py>=2.3.0

# HTTP/WebSocket
aiohttp>=3.9.0
websockets>=12.0

# Configuration
PyYAML>=6.0

# Timezone
pytz>=2024.1
